fix get options

+++ programs/witness_node/main.cpp
@@ -249,7 +249,7 @@ int main(int argc, char** argv)
       bpo::notify(options);
       
       auto _options = &options;
-      auto max_market = _options->at("max-order-his-seconds-per-market").as<uint64_t>();
+      auto max_market = _options->at("max-order-his-seconds-per-market").as<uint32_t>();
       auto contract_private_data_size = _options->at("contract_private_data_size").as<uint64_t>();
       wlog("------------------------------------------------------- ${test}",("test", max_market));
       wlog("------------------------------------------------------- ${test}",("test", contract_private_data_size));
